create helpCentreSingleTopic.tsx
